+++ library/src/main/cpp/util.cpp
@@ -66,11 +66,12 @@ namespace util {
                               );
     }
 
-    std::pair<jstring, jbyteArray> deserialize_user_pic(JNIEnv *env, jobject user_pic) {
-        auto userPicClass = jni_utils::JavaLocalRef(env, env->GetObjectClass(user_pic));
+
+    session::config::profile_pic deserialize_user_pic(JNIEnv *env, jobject user_pic) {
+        jni_utils::JavaLocalRef clazz(env, env->GetObjectClass(user_pic));
         return {
-            static_cast<jstring>(env->CallObjectMethod(user_pic, env->GetMethodID(userPicClass.get(), "getUrl", "()Ljava/lang/String;"))),
-            static_cast<jbyteArray>(env->CallObjectMethod(user_pic, env->GetMethodID(userPicClass.get(), "getKeyAsByteArray", "()[B")))
+            jni_utils::JavaStringRef(env, (jstring) (env->CallObjectMethod(user_pic, env->GetMethodID(clazz.get(), "getUrl", "Ljava/lang/String;")))).view(),
+            util::vector_from_bytes(env, static_cast<jbyteArray>(env->CallObjectMethod(user_pic, env->GetMethodID(clazz.get(), "getKeyAsByteArray", "()[B"))))
         };
     }
