+++ library/src/main/cpp/util.h
@@ -21,7 +21,7 @@ namespace util {
     jbyteArray bytes_from_span(JNIEnv* env, std::span<const unsigned char> from_str);
     std::string string_from_jstring(JNIEnv* env, jstring string);
     jobject serialize_user_pic(JNIEnv *env, session::config::profile_pic pic);
-    std::pair<jstring, jbyteArray> deserialize_user_pic(JNIEnv *env, jobject user_pic);
+    session::config::profile_pic deserialize_user_pic(JNIEnv *env, jobject user_pic);
     jobject serialize_base_community(JNIEnv *env, const session::config::community& base_community);
     session::config::community deserialize_base_community(JNIEnv *env, jobject base_community);
     jobject serialize_expiry(JNIEnv *env, const session::config::expiration_mode& mode, const std::chrono::seconds& time_seconds);
