+++ library/src/main/java/network/loki/messenger/libsession_util/util/BaseCommunity.kt
@@ -7,5 +7,6 @@ data class BaseCommunityInfo(val baseUrl: String, val room: String, val pubKeyHe
         }
         external fun parseFullUrl(fullUrl: String): Triple<String, String, ByteArray>?
     }
+
     external fun fullUrl(): String
 }
