+++ library/src/main/java/network/loki/messenger/libsession_util/util/Contact.kt
@@ -8,6 +8,8 @@ data class Contact(
     var approvedMe: Boolean = false,
     var blocked: Boolean = false,
     var profilePicture: UserPic = UserPic.DEFAULT,
+    var createdEpochSeconds: Long = 0,
+    var profileUpdatedEpochSeconds: Long = 0,
     var priority: Long = 0,
     var expiryMode: ExpiryMode = ExpiryMode.NONE,
 ) {
